+++ src/main/java/org/betonquest/betonquest/quest/event/hunger/HungerEvent.java
@@ -1,5 +1,6 @@
 package org.betonquest.betonquest.quest.event.hunger;
 
+import org.betonquest.betonquest.VariableNumber;
 import org.betonquest.betonquest.api.profiles.Profile;
 import org.betonquest.betonquest.api.quest.event.Event;
 import org.betonquest.betonquest.exceptions.QuestRuntimeException;
@@ -17,7 +18,7 @@ public class HungerEvent implements Event {
     /**
      * The amount of hunger to apply.
      */
-    private final int amount;
+    private final VariableNumber amount;
 
     /**
      * Create the hunger event to set the given state.
@@ -25,7 +26,7 @@ public class HungerEvent implements Event {
      * @param hunger the hunger type
      * @param amount the amount of hunger to apply
      */
-    public HungerEvent(final Hunger hunger, final int amount) {
+    public HungerEvent(final Hunger hunger, final VariableNumber amount) {
         this.hunger = hunger;
         this.amount = amount;
     }
@@ -33,6 +34,6 @@ public class HungerEvent implements Event {
     @Override
     public void execute(final Profile profile) throws QuestRuntimeException {
         final Player player = profile.getOnlineProfile().get().getPlayer();
-        player.setFoodLevel(hunger.calculate(player, amount));
+        player.setFoodLevel(hunger.calculate(player, amount.getInt(profile)));
     }
 }